refactor: move Tools enum to types.ts
